test(memory): probe tests against real endpoints, not asserted mock kwargs
CI failed every test lane on the H1 fix. The cause was the local
verification command, not the fix: this repo keeps memory tests in TWO
trees — tests/unit/memory/ AND tests/memory/ — and only the first was
run. testpaths = ["tests"] is what CI honours.

The two failing tests are class M in its purest form.
tests/memory/test_redis_bootstrap.py::TestCheckRedisRunning patched
redis.Redis and asserted the kwargs the mock was called with, which
proves the test handed the mock what the test intended and says nothing
about which endpoint gets dialled. That is exactly how H1 — a probe
hard-coded to 127.0.0.1:6379 while clients connect to the resolved
endpoint — survived to production.

Rewritten against real sockets: an endpoint that answers, an endpoint
with nothing listening, and an explicit host/port overriding the
resolved one in BOTH directions. Four kwargs-asserting tests become
three behavioural ones.

The RESP stub moves to tests/support/redis_stub.py so the H1 suite and
this one share one real-endpoint helper instead of duplicating it.

Receipts: tests/memory + tests/unit/memory + tests/unit/gates —
3044 passed. The 5 remaining failures are
tests/memory/test_ams_backend_integration.py, pre-existing on clean main
(a local agent-memory-server returning HTTP 500) and skipped in CI.
