+++ tests/support/__init__.py
@@ -1,0 +1,1 @@
+"""Shared test support helpers (not tests themselves)."""
